requirements.txt
Wand==0.6.11  # Python binding for ImageMagick
Pillow==10.2.0  # Additional image processing capabilities 